# perf: faster Payroll  :- Part 1 (Payroll Enty)