--- a/hrms/payroll/doctype/employee_benefit_application/test_employee_benefit_application.py
+++ b/hrms/payroll/doctype/employee_benefit_application/test_employee_benefit_application.py
@@ -55,6 +55,7 @@ class TestEmployeeBenefitApplication(FrappeTestCase):
 			include_flexi_benefits=True,
 			employee=employee,
 			payroll_period=payroll_period,
+			company="_Test Company",
 		)
 		salary_slip = make_salary_slip(salary_structure.name, employee=employee, posting_date=getdate())
 		salary_slip.insert()