--- a/hrms/payroll/doctype/payroll_entry/payroll_entry.js
+++ b/hrms/payroll/doctype/payroll_entry/payroll_entry.js
@@ -393,7 +393,7 @@ let make_bank_entry = function (frm) {
 		return frappe.call({
 			method: "run_doc_method",
 			args: {
-				method: "make_payment_entry",
+				method: "make_bank_entry",
 				dt: "Payroll Entry",
 				dn: frm.doc.name
 			},