… hrms/payroll/doctype/payroll_entry/test_payroll_entry.py
… hrms/payroll/doctype/salary_slip/salary_slip.py
… hrms/payroll/doctype/salary_structure/salary_structure.json