Tested on 100k Employee records

## Employee Selections
#### Before:
<img width="1434" alt="Screenshot 2023-04-03 at 4 31 41 PM" src="https://user-images.githubusercontent.com/3784093/229491930-ad9bc14c-4fca-4bb0-82d6-20817483c15c.png">

search_link: apx 2.23 sec

#### After:

<img width="1440" alt="Screenshot 2023-04-03 at 4 33 45 PM" src="https://user-images.githubusercontent.com/3784093/229492227-10891f69-ad3e-4141-b5e1-dd8e81bd17b6.png">

search_link: apx 62 ms

---
Data Chunk 7k employees

## Fill Employee table based on filters
### Before
<img width="1432" alt="Screenshot 2023-04-03 at 4 32 43 PM" src="https://user-images.githubusercontent.com/3784093/229491957-0fae2f3e-8f4e-4fb7-9ff2-25999122fff0.png">

run_doc_method = 12.14
savedoc = 19.29

### After
<img width="1433" alt="Screenshot 2023-04-03 at 4 34 48 PM" src="https://user-images.githubusercontent.com/3784093/229492384-06283d7d-a138-4dad-b468-7abb8990ab37.png">


run_doc_method = 10.16 sec
savedoc = 19.37 sec

---

## validate salary slip against employees
### Before

<img width="701" alt="Screenshot 2023-04-10 at 3 29 14 PM" src="https://user-images.githubusercontent.com/3784093/230880350-e8e16ef9-94f5-4e17-ab0c-9659f90b9507.png">

Time Required = 7 Sec

### After
<img width="764" alt="Screenshot 2023-04-10 at 3 29 41 PM" src="https://user-images.githubusercontent.com/3784093/230880397-9b9af4d9-4967-47ce-8fb4-692fcae47e79.png">

Time Required = 0.166

---
## Validate employee attendance

### Before
<img width="728" alt="Screenshot 2023-04-10 at 5 24 55 PM" src="https://user-images.githubusercontent.com/3784093/230899293-2e84faec-711e-41f9-9367-777f890318bb.png">

### After
<img width="806" alt="Screenshot 2023-04-10 at 5 40 22 PM" src="https://user-images.githubusercontent.com/3784093/230899323-7d6313f5-ec36-48fd-a701-0e449a95a960.png">

---
## make_payment_entry

### Before
<img width="630" alt="Screenshot 2023-04-12 at 12 54 11 PM" src="https://user-images.githubusercontent.com/3784093/231384396-a436decb-fb43-40bc-aca1-add7959555d5.png">

### After
<img width="700" alt="Screenshot 2023-04-12 at 12 55 40 PM" src="https://user-images.githubusercontent.com/3784093/231384444-e21b3a54-6e39-448b-94e0-f7c67665088b.png">