@ruchamahabal 2023-04-24
Better commit conventions can be used:

`perf:` for performance fixes
`refactor:` for such commits instead of `chore:`: [`04a1908` (#417)](https://github.com/frappe/hrms/pull/417/commits/04a190841174382b590cd9f65cf7b302d981ca8d)