@ruchamahabal 2023-05-09
reverting "fix: handle cancellation via background job for higher employees count" 

Reasons:
- Queuing salary slip deletion needs more handling and UX work. The current one doesn’t function correctly: status changes to cancelled immediately but it shows as queued
- Also if salary slip is submitted, deletion fails. ignoring gl entries during deletion is also not functioning well

https://user-images.githubusercontent.com/24353136/237057318-798d91be-e3ce-44e7-9ad1-0918ceb31ea9.mp4

Will be taken care of in a separate PR